@@ -22,13 +22,15 @@
     "tsc": "lerna run tsc --no-private"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/preset-typescript": "^7.3.3",
-    "@types/express": "^4.17.0",
+    "@babel/core": "^7.6.4",
+    "@babel/preset-typescript": "^7.6.0",
+    "@types/cross-spawn": "^6.0.1",
+    "@types/express": "^4.17.1",
     "@types/fs-extra": "8.0.0",
     "@types/inquirer": "^6.0.3",
     "@types/jest": "^24.0.15",
-    "@types/lodash": "^4.14.136",
+    "@types/loader-utils": "^1.1.3",
+    "@types/lodash": "^4.14.144",
     "@types/lodash.kebabcase": "^4.1.6",
     "@types/node": "^12.0.2",
     "@types/react": "^16.8.18",
@@ -53,19 +55,23 @@
     "front-matter": "^2.3.0",
     "glob-promise": "^3.3.0",
     "husky": "^1.3.1",
-    "jest": "^24.6.0",
-    "lerna": "^3.14.1",
+    "jest": "^24.9.0",
+    "lerna": "^3.18.1",
     "lint-staged": "^7.2.0",
-    "prettier": "^1.13.7",
+    "prettier": "^1.18.2",
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
     "rimraf": "^2.6.3",
-    "typescript": "^3.5.3"
+    "typescript": "^3.6.4"
   },
   "lint-staged": {
     "linters": {
-      "*.js": [
+      "*.{js,jsx}": [
         "yarn lint --fix",
+        "yarn prettier",
+        "git add"
+      ],
+      "*.{ts,tsx}": [
         "yarn prettier",
         "git add"
       ]
