@@ -1,7 +1,7 @@
 {
   "name": "@docusaurus/core",
   "description": "Easy to Maintain Open Source Documentation Websites",
-  "version": "2.0.0-alpha.24",
+  "version": "2.0.0-alpha.31",
   "license": "MIT",
   "publishConfig": {
     "access": "public"
@@ -17,7 +17,7 @@
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/facebook/Docusaurus.git"
+    "url": "https://github.com/facebook/docusaurus.git"
   },
   "bin": {
     "docusaurus": "bin/docusaurus.js"
@@ -26,57 +26,61 @@
     "tsc": "tsc && node copyUntypedFiles.js"
   },
   "bugs": {
-    "url": "https://github.com/facebook/Docusaurus/issues"
+    "url": "https://github.com/facebook/docusaurus/issues"
+  },
+  "devDependencies": {
+    "@docusaurus/types": "^2.0.0-alpha.31"
   },
   "dependencies": {
-    "@babel/core": "^7.5.5",
+    "@babel/core": "^7.6.4",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-react": "^7.0.0",
-    "@docusaurus/utils": "^2.0.0-alpha.24",
+    "@babel/preset-env": "^7.6.3",
+    "@babel/preset-react": "^7.6.3",
+    "@docusaurus/utils": "^2.0.0-alpha.31",
     "babel-loader": "^8.0.6",
     "babel-plugin-dynamic-import-node": "^2.3.0",
     "cache-loader": "^4.1.0",
     "chalk": "^2.4.2",
-    "chokidar": "^3.0.2",
+    "chokidar": "^3.2.2",
     "classnames": "^2.2.6",
-    "clean-webpack-plugin": "^2.0.1",
     "commander": "^2.20.0",
     "copy-webpack-plugin": "^5.0.4",
-    "css-loader": "^3.1.0",
+    "core-js": "^2.6.5",
+    "css-loader": "^3.2.0",
+    "del": "^4.1.1",
     "ejs": "^2.6.2",
     "express": "^4.17.1",
     "fs-extra": "^8.1.0",
     "globby": "^10.0.1",
-    "html-webpack-plugin": "^4.0.0-beta.7",
-    "import-fresh": "^3.0.0",
+    "html-webpack-plugin": "^4.0.0-beta.8",
+    "import-fresh": "^3.1.0",
     "lodash": "^4.17.15",
     "mini-css-extract-plugin": "^0.8.0",
     "nprogress": "^0.2.0",
     "null-loader": "^3.0.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "portfinder": "^1.0.21",
+    "portfinder": "^1.0.25",
     "postcss-loader": "^3.0.0",
     "postcss-preset-env": "^6.7.0",
-    "react-dev-utils": "^9.0.1",
+    "react-dev-utils": "^9.1.0",
     "react-helmet": "^6.0.0-beta",
     "react-loadable": "^5.5.0",
-    "react-loadable-ssr-addon": "^0.1.9",
-    "react-router": "^5.0.1",
-    "react-router-config": "^5.0.1",
-    "react-router-dom": "^5.0.1",
+    "react-loadable-ssr-addon": "^0.2.0",
+    "react-router": "^5.1.2",
+    "react-router-config": "^5.1.1",
+    "react-router-dom": "^5.1.2",
     "semver": "^6.3.0",
     "shelljs": "^0.8.3",
-    "static-site-generator-webpack-plugin": "^3.4.2",
+    "@endiliey/static-site-generator-webpack-plugin": "^4.0.0",
     "std-env": "^2.2.1",
-    "style-loader": "^0.23.1",
-    "terser-webpack-plugin": "^1.3.0",
-    "wait-file": "^1.0.3",
-    "webpack": "^4.38.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-dev-server": "^3.7.2",
-    "webpack-merge": "^4.2.1",
-    "webpack-nicelog": "^2.3.1"
+    "style-loader": "^1.0.0",
+    "terser-webpack-plugin": "^2.1.3",
+    "wait-file": "^1.0.5",
+    "webpack": "^4.41.2",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-dev-server": "^3.8.2",
+    "webpack-merge": "^4.2.2",
+    "webpackbar": "^4.0.0"
   },
   "peerDependencies": {
     "react": "^16.8.4",
