@@ -7,17 +7,23 @@
 
 const path = require('path');
 
+const ignorePatterns = [
+  '/node_modules/',
+  '__fixtures__',
+  '/packages/docusaurus/lib',
+  '/packages/docusaurus-utils/lib',
+  '/packages/docusaurus-plugin-content-blog/lib',
+  '/packages/docusaurus-plugin-content-docs/lib',
+  '/packages/docusaurus-plugin-content-pages/lib',
+];
+
 module.exports = {
   rootDir: path.resolve(__dirname),
   verbose: true,
   testURL: 'http://localhost/',
   testEnvironment: 'node',
-  testPathIgnorePatterns: [
-    '/node_modules/',
-    '__fixtures__',
-    '/packages/docusaurus/lib',
-    '/packages/docusaurus-utils/lib',
-  ],
+  testPathIgnorePatterns: ignorePatterns,
+  coveragePathIgnorePatterns: ignorePatterns,
   transform: {
     '^.+\\.[jt]sx?$': 'babel-jest',
   },
