@@ -174,9 +174,7 @@ class HeaderNav extends React.Component {
           : '';
       const id = langPart + versionPart + link.doc;
       if (!Metadata[id]) {
-        let errorStr = `Processing the following \`doc\` field in \`headerLinks\` within \`siteConfig.js\`: '${
-          link.doc
-        }'`;
+        let errorStr = `Processing the following \`doc\` field in \`headerLinks\` within \`siteConfig.js\`: '${link.doc}'`;
         if (id === link.doc) {
           errorStr +=
             ' It looks like there is no document with that id that exists in your docs directory. Please double check the spelling of your `doc` field and the `id` fields of your docs.';
