@@ -7,6 +7,7 @@
 
 const Remarkable = require('remarkable');
 const mdToc = require('markdown-toc');
+const striptags = require('striptags');
 const toSlug = require('./toSlug');
 
 const tocRegex = new RegExp('<AUTOGENERATED_TABLE_OF_CONTENTS>', 'i');
@@ -34,15 +35,20 @@ function getTOC(content, headingTags = 'h2', subHeadingTags = 'h3') {
   headings.forEach(heading => {
     // we need always generate slugs to ensure, that we will have consistent
     // slug indexes for headings with the same names
-    const hashLink = toSlug(heading.content, context);
+    const rawContent = heading.content;
+    const safeContent = striptags(rawContent);
+    const rendered = md.renderInline(safeContent);
+
+    // We striptags again here as to not end up with html tags
+    // from markdown or markdown in our links
+    const hashLink = toSlug(striptags(rendered), context);
     if (!allowedHeadingLevels.includes(heading.lvl)) {
       return;
     }
-    const rawContent = heading.content;
     const entry = {
       hashLink,
       rawContent,
-      content: md.renderInline(rawContent),
+      content: rendered,
       children: [],
     };
     if (headingLevels.includes(heading.lvl)) {
