@@ -191,13 +191,9 @@ function processVersionMetadata(file, version, useVersion, language) {
   const docsPart = `${siteConfig.docsUrl ? `${siteConfig.docsUrl}/` : ''}`;
   const versionPart = `${version !== latestVersion ? `${version}/` : ''}`;
   if (!ENABLE_TRANSLATION && !siteConfig.useEnglishUrl) {
-    metadata.permalink = `${docsPart}${versionPart}${
-      metadata.original_id
-    }.html`;
+    metadata.permalink = `${docsPart}${versionPart}${metadata.original_id}.html`;
   } else {
-    metadata.permalink = `${docsPart}${language}/${versionPart}${
-      metadata.original_id
-    }.html`;
+    metadata.permalink = `${docsPart}${language}/${versionPart}${metadata.original_id}.html`;
   }
   metadata.id = metadata.id.replace(
     `version-${useVersion}-`,
