@@ -10,46 +10,41 @@ import React from 'react';
 import Layout from '@theme/Layout';
 import Link from '@docusaurus/Link';
 
-const CHARS_IN_ALPHABET = 26;
-const ASCII_LOWERCASE_A = 97;
+function getCategoryOfTag(tag) {
+  // tag's category should be customizable
+  return tag[0].toUpperCase();
+}
 
 function BlogTagsListPage(props) {
   const {tags} = props;
 
-  const tagsList = Array(CHARS_IN_ALPHABET)
-    .fill(null)
-    .map(() => []);
-  const allTags = Object.keys(tags).sort();
-
-  allTags.forEach(tag => {
-    const firstLetter = tag.charCodeAt(0);
-    tagsList[firstLetter - ASCII_LOWERCASE_A].push(tag);
+  const tagCategories = {};
+  Object.keys(tags).forEach(tag => {
+    const category = getCategoryOfTag(tag);
+    tagCategories[category] = tagCategories[category] || [];
+    tagCategories[category].push(tag);
   });
-
+  const tagsList = Object.entries(tagCategories).sort(([a], [b]) => {
+    if (a === b) {
+      return 0;
+    }
+    return a > b ? 1 : -1;
+  });
   const tagsSection = tagsList
-    .map((tagsForLetter, index) => {
-      if (tagsForLetter.length === 0) {
-        return null;
-      }
-      const letter = String.fromCharCode(
-        ASCII_LOWERCASE_A + index,
-      ).toUpperCase();
-
-      return (
-        <div key={letter}>
-          <h3>{letter}</h3>
-          {tagsForLetter.map(tag => (
-            <Link
-              className="padding-right--md"
-              href={tags[tag].permalink}
-              key="tag">
-              {tags[tag].name} ({tags[tag].count})
-            </Link>
-          ))}
-          <hr />
-        </div>
-      );
-    })
+    .map(([category, tagsForCategory]) => (
+      <div key={category}>
+        <h3>{category}</h3>
+        {tagsForCategory.map(tag => (
+          <Link
+            className="padding-right--md"
+            href={tags[tag].permalink}
+            key={tag}>
+            {tags[tag].name} ({tags[tag].count})
+          </Link>
+        ))}
+        <hr />
+      </div>
+    ))
     .filter(item => item != null);
 
   return (
