@@ -101,19 +101,27 @@ export function genChunkName(
   modulePath: string,
   prefix?: string,
   preferredName?: string,
+  shortId?: boolean,
 ): string {
   let chunkName: string | undefined = chunkNameCache.get(modulePath);
   if (!chunkName) {
-    let str = modulePath;
-    if (preferredName) {
-      const shortHash = createHash('md5')
+    if (shortId) {
+      chunkName = createHash('md5')
         .update(modulePath)
         .digest('hex')
-        .substr(0, 3);
-      str = `${preferredName}${shortHash}`;
+        .substr(0, 8);
+    } else {
+      let str = modulePath;
+      if (preferredName) {
+        const shortHash = createHash('md5')
+          .update(modulePath)
+          .digest('hex')
+          .substr(0, 3);
+        str = `${preferredName}${shortHash}`;
+      }
+      const name = str === '/' ? 'index' : docuHash(str);
+      chunkName = prefix ? `${prefix}---${name}` : name;
     }
-    const name = str === '/' ? 'index' : docuHash(str);
-    chunkName = prefix ? `${prefix}---${name}` : name;
     chunkNameCache.set(modulePath, chunkName);
   }
   return chunkName;
