@@ -12,7 +12,7 @@ import path from 'path';
 import TerserPlugin from 'terser-webpack-plugin';
 import {Configuration, Loader} from 'webpack';
 
-import {Props} from '../server/types';
+import {Props} from '@docusaurus/types';
 import {getBabelLoader, getCacheLoader, getStyleLoaders} from './utils';
 
 const CSS_REGEX = /\.css$/;
@@ -22,15 +22,9 @@ export function createBaseConfig(
   props: Props,
   isServer: boolean,
 ): Configuration {
-  const {
-    outDir,
-    siteDir,
-    baseUrl,
-    generatedFilesDir,
-    cliOptions: {cacheLoader},
-    routesPaths,
-  } = props;
+  const {outDir, siteDir, baseUrl, generatedFilesDir, routesPaths} = props;
 
+  const clientDir = path.join(__dirname, '..', 'client');
   const totalPages = routesPaths.length;
   const isProd = process.env.NODE_ENV === 'production';
   return {
@@ -38,8 +32,8 @@ export function createBaseConfig(
     output: {
       pathinfo: false,
       path: outDir,
-      filename: isProd ? '[name].[contenthash].js' : '[name].js',
-      chunkFilename: isProd ? '[name].[contenthash].js' : '[name].js',
+      filename: isProd ? '[name].[contenthash:8].js' : '[name].js',
+      chunkFilename: isProd ? '[name].[contenthash:8].js' : '[name].js',
       publicPath: baseUrl,
     },
     // Don't throw warning when asset created is over 250kb
@@ -56,7 +50,12 @@ export function createBaseConfig(
         '@generated': generatedFilesDir,
         '@docusaurus': path.resolve(__dirname, '../client/exports'),
       },
+      // This allows you to set a fallback for where Webpack should look for modules.
+      // We want `@docusaurus/core` own dependencies/`node_modules` to "win" if there is conflict
+      // Example: if there is core-js@3 in user's own node_modules, but core depends on
+      // core-js@2, we should use core-js@2.
       modules: [
+        path.resolve(__dirname, '..', '..', 'node_modules'),
         'node_modules',
         path.resolve(fs.realpathSync(process.cwd()), 'node_modules'),
       ],
@@ -72,10 +71,27 @@ export function createBaseConfig(
               parallel: true,
               sourceMap: false,
               terserOptions: {
-                ecma: 6,
-                mangle: true,
+                parse: {
+                  // we want uglify-js to parse ecma 8 code. However, we don't want it
+                  // to apply any minfication steps that turns valid ecma 5 code
+                  // into invalid ecma 5 code. This is why the 'compress' and 'output'
+                  // sections only apply transformations that are ecma 5 safe
+                  // https://github.com/facebook/create-react-app/pull/4234
+                  ecma: 8,
+                },
+                compress: {
+                  ecma: 5,
+                  warnings: false,
+                },
+                mangle: {
+                  safari10: true,
+                },
                 output: {
+                  ecma: 5,
                   comments: false,
+                  // Turned on because emoji and regex is not minified properly using default
+                  // https://github.com/facebook/create-react-app/issues/2488
+                  ascii_only: true,
                 },
               },
             }),
@@ -105,15 +121,19 @@ export function createBaseConfig(
         {
           test: /\.jsx?$/,
           exclude: modulePath => {
-            // Don't transpile node_modules except any docusaurus package
+            // always transpile client dir
+            if (modulePath.startsWith(clientDir)) {
+              return false;
+            }
+            // Don't transpile node_modules except any docusaurus npm package
             return (
-              /node_modules/.test(modulePath) && !/docusaurus/.test(modulePath)
+              /node_modules/.test(modulePath) &&
+              !/(docusaurus)((?!node_modules).)*\.jsx?$/.test(modulePath)
             );
           },
-          use: [
-            cacheLoader && getCacheLoader(isServer),
-            getBabelLoader(isServer),
-          ].filter(Boolean) as Loader[],
+          use: [getCacheLoader(isServer), getBabelLoader(isServer)].filter(
+            Boolean,
+          ) as Loader[],
         },
         {
           test: CSS_REGEX,
@@ -129,7 +149,9 @@ export function createBaseConfig(
           test: CSS_MODULE_REGEX,
           use: getStyleLoaders(isServer, {
             modules: {
-              localIdentName: `[local]_[hash:base64:4]`,
+              localIdentName: isProd
+                ? `[local]_[hash:base64:4]`
+                : `[local]_[path]`,
             },
             importLoaders: 1,
             sourceMap: !isProd,
@@ -140,8 +162,8 @@ export function createBaseConfig(
     },
     plugins: [
       new MiniCssExtractPlugin({
-        filename: isProd ? '[name].[contenthash].css' : '[name].css',
-        chunkFilename: isProd ? '[name].[contenthash].css' : '[name].css',
+        filename: isProd ? '[name].[contenthash:8].css' : '[name].css',
+        chunkFilename: isProd ? '[name].[contenthash:8].css' : '[name].css',
       }),
     ],
   };
