@@ -54,6 +54,7 @@ export function getStyleLoaders(
             autoprefixer: {
               flexbox: 'no-2009',
             },
+            stage: 3,
           }),
         ],
       },
@@ -88,7 +89,30 @@ export function getBabelLoader(isServer: boolean, babelOptions?: {}): Loader {
       {
         babelrc: false,
         configFile: false,
-        presets: ['@babel/env', '@babel/react'],
+        presets: [
+          isServer
+            ? [
+                '@babel/env',
+                {
+                  targets: {
+                    node: 'current',
+                  },
+                },
+              ]
+            : [
+                '@babel/env',
+                {
+                  useBuiltIns: 'usage',
+                  loose: true,
+                  corejs: '2',
+                  // Do not transform modules to CJS
+                  modules: false,
+                  // Exclude transforms that make all code slower
+                  exclude: ['transform-typeof-symbol'],
+                },
+              ],
+          '@babel/react',
+        ],
         plugins: [
           isServer ? 'dynamic-import-node' : '@babel/syntax-dynamic-import',
         ],
