@@ -11,8 +11,8 @@ const users = [
     title: 'Channel.js',
     description: 'The missing constructor for creating safe async iterators',
     preview: require('./showcase/channeljs.png'),
-    website: 'https://channel.js.org',
-    source: 'https://github.com/channeljs/channel',
+    website: 'https://repeater.js.org/',
+    source: 'https://github.com/repeaterjs/repeater',
     fbOpenSource: false,
     pinned: false,
   },
@@ -42,6 +42,15 @@ const users = [
     source: 'https://github.com/facebook/hermes',
     fbOpenSource: true,
     pinned: true,
+  },
+  {
+    title: 'MBT Bundle',
+    description: 'An open source model-based testing tool',
+    preview: require('./showcase/mbt-bundle.png'),
+    website: 'https://mbtbundle.org',
+    source: 'https://github.com/tienvx/mbt-bundle',
+    fbOpenSource: false,
+    pinned: false,
   },
   {
     title: 'uniforms',
