@@ -9,6 +9,8 @@ import React, {useEffect} from 'react';
 import Layout from '@theme/Layout';
 
 import canny from '../../scripts/canny';
+import classnames from 'classnames';
+import styles from './styles.module.css';
 
 const BOARD_TOKEN = '054e0e53-d951-b14c-7e74-9eb8f9ed2f91';
 
@@ -24,7 +26,14 @@ function Feedback() {
 
   return (
     <Layout permalink={'/feedback'} description={'Docusaurus 2 Feedback page'}>
-      <div className="container margin-vert--xl" data-canny />
+      <div
+        className={classnames(
+          'container',
+          'margin-vert--xl',
+          styles.feedbackBackground,
+        )}
+        data-canny
+      />
     </Layout>
   );
 }
